refactor: isolated prefetch queries
